feat: DriveViewModelで統一した
