+++ modules/features/drive/src/main/java/net/pantasystem/milktea/drive/CreateFolderDialog.kt
@@ -6,7 +6,7 @@ import android.view.View
 import androidx.appcompat.app.AppCompatDialogFragment
 import androidx.lifecycle.ViewModelProvider
 import net.pantasystem.milktea.drive.databinding.DialogCreateFolderBinding
-import net.pantasystem.milktea.drive.viewmodel.DirectoryViewModel
+import net.pantasystem.milktea.drive.viewmodel.DriveViewModel
 
 class CreateFolderDialog : AppCompatDialogFragment(){
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
@@ -14,7 +14,7 @@ class CreateFolderDialog : AppCompatDialogFragment(){
         val view = View.inflate(dialog.context, R.layout.dialog_create_folder, null)
         val binding = DialogCreateFolderBinding.bind(view)
         dialog.setContentView(view)
-        val directoryViewModel = ViewModelProvider(requireActivity())[DirectoryViewModel::class.java]
+        val directoryViewModel = ViewModelProvider(requireActivity())[DriveViewModel::class.java]
         binding.okButton.setOnClickListener {
             val name = binding.editFolderName.text.toString()
             if(name.isNotBlank()){
