--- modules/features/drive/src/main/java/net/pantasystem/milktea/drive/viewmodel/DirectoryViewData.kt
@@ -1,12 +1,0 @@
-package net.pantasystem.milktea.drive.viewmodel
-
-import net.pantasystem.milktea.model.drive.Directory
-
-data class DirectoryViewData (val directory: Directory){
-    val id = directory.id
-    val createdAt = directory.createdAt
-    val name = directory.name
-    val parent = directory.parent
-
-
-}
